rework mapping execution builder and mapping test editor
