+++ packages/legend-studio-plugin-query-builder/style/_mapping-execution-query-builder.scss
@@ -16,7 +16,7 @@
 
 .mapping-execution-query-builder {
   &__btn {
-    height: 2.6rem;
-    width: calc(100% - 1rem);
+    height: 2.8rem;
+    width: 100%;
   }
 }
