+++ packages/legend-studio-plugin-query-builder/style/_mapping-test-query-builder.scss
@@ -16,7 +16,7 @@
 
 .mapping-test-query-builder {
   &__btn {
-    height: 2.6rem;
-    width: calc(100% - 1rem);
+    height: 2.8rem;
+    width: 100%;
   }
 }
