+++ packages/legend-studio-plugin-query-builder/style/_service-query-builder.scss
@@ -1,0 +1,8 @@
+.service-query-builder {
+  padding: 1rem;
+
+  &__btn {
+    width: 100%;
+    height: 2.8rem;
+  }
+}
