+++ packages/legend-studio-plugin-query-builder/style/index.scss
@@ -15,6 +15,7 @@
  */
 
 @forward 'query-builder';
+@forward 'service-query-builder';
 @forward 'mapping-execution-query-builder';
 @forward 'mapping-test-query-builder';
 @forward 'graph-fetch-tree';
