--- packages/legend-studio/style/components/editor/aux-panel/_mapping-execution.scss
@@ -1,123 +1,0 @@
-/**
- * Copyright (c) 2020-present, Goldman Sachs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@use '../../../mixins' as *;
-
-.mapping-execution-panel {
-  height: 100%;
-  width: 100%;
-  position: relative;
-
-  &--no-content {
-    padding: 2rem;
-  }
-
-  &--empty {
-    @include flexCenter;
-
-    height: 100%;
-    width: 100%;
-    text-align: center;
-    line-height: 1.8rem;
-    color: var(--color-dark-grey-500);
-  }
-
-  .panel {
-    height: 100%;
-
-    &__header {
-      @include flexVCenter;
-
-      color: var(--color-light-grey-400);
-      font-weight: bold;
-      background: var(--color-dark-grey-50);
-      border-top: 0.1rem solid var(--color-dark-grey-100);
-    }
-
-    &__header__title__label {
-      color: var(--color-dark-grey-50);
-      background: var(--color-light-grey-300);
-    }
-
-    &__header__actions {
-      margin-right: 0.5rem;
-    }
-
-    &__header__action svg {
-      color: var(--color-light-grey-400);
-    }
-
-    &__header__action[disabled] svg {
-      color: var(--color-dark-grey-300) !important;
-    }
-
-    &__content {
-      background: var(--color-dark-grey-50);
-    }
-
-    &__content--dnd-over {
-      filter: brightness(105%) blur(0.05rem);
-      border: 0.1rem solid var(--color-light-blue-200);
-    }
-  }
-
-  &__query-panel {
-    height: 100%;
-  }
-
-  &__query-panel__query {
-    position: relative;
-    height: calc(100% - 3.6rem);
-  }
-
-  &__query-panel__query-editor {
-    @include flexCenter;
-
-    height: 3.6rem;
-    border-top: 0.1rem solid var(--color-dark-grey-200);
-  }
-
-  .Resizer.vertical {
-    background: padding-box var(--color-dark-grey-100);
-    width: 1.1rem;
-    min-width: 1.1rem;
-    border-left: 0.5rem solid transparent;
-    border-right: 0.5rem solid transparent;
-  }
-
-  &__input-data-panel__content {
-    padding: 0 !important;
-    position: relative;
-  }
-
-  &__result-panel__content {
-    position: relative;
-  }
-
-  &__generate-plan-btn svg {
-    font-size: 1.5rem;
-  }
-
-  &__service__modal {
-    &__service-path {
-      width: 100%;
-      background: var(--color-dark-grey-280);
-      color: var(--color-light-grey-200);
-      border: 0.1rem solid var(--color-dark-grey-300);
-      margin-bottom: 1rem;
-    }
-  }
-}
