--- packages/legend-studio/style/components/editor/aux-panel/_mapping-test-editor-panel.scss
@@ -1,147 +1,0 @@
-/**
- * Copyright (c) 2020-present, Goldman Sachs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@use '../../../mixins' as *;
-
-.mapping-test-editor-panel {
-  height: 100%;
-  width: 100%;
-  position: relative;
-
-  &--no-content {
-    padding: 2rem;
-  }
-
-  .panel {
-    height: 100%;
-
-    &__header {
-      @include flexVCenter;
-
-      color: var(--color-light-grey-400);
-      font-weight: bold;
-      background: var(--color-dark-grey-50);
-      border-top: 0.1rem solid var(--color-dark-grey-100);
-    }
-
-    &__header__title__label {
-      color: var(--color-dark-grey-50);
-      background: var(--color-light-grey-300);
-    }
-
-    &__header__actions {
-      margin-right: 0.5rem;
-    }
-
-    &__header__action svg {
-      color: var(--color-light-grey-400);
-    }
-
-    &__header__action[disabled] svg {
-      color: var(--color-dark-grey-300) !important;
-    }
-
-    &__header__action.btn--dark svg {
-      color: inherit;
-    }
-
-    &__content {
-      background: var(--color-dark-grey-50);
-    }
-
-    &__content--dnd-over {
-      filter: brightness(105%) blur(0.05rem);
-      border: 0.1rem solid var(--color-light-blue-200);
-    }
-  }
-
-  &__target-panel .panel__content {
-    color: var(--color-light-grey-400);
-  }
-
-  &__target-panel__query-container {
-    height: 100%;
-    width: 100%;
-    padding: 0.5rem 0;
-    overflow: auto;
-  }
-
-  &__target {
-    @include flexCenter;
-
-    height: 100%;
-    flex-direction: column;
-    cursor: pointer;
-    margin-top: 0 !important;
-  }
-
-  &__target__text {
-    @include flexHCenter;
-
-    color: var(--color-dark-grey-400);
-    font-weight: bold;
-  }
-
-  &__target__action {
-    @include flexCenter;
-
-    padding: 1rem;
-    border: 0.3rem dashed var(--color-dark-grey-100);
-    border-radius: 0.3rem;
-    margin-top: 2rem;
-    width: 10rem;
-    height: 10rem;
-  }
-
-  &__target__action svg {
-    font-size: 4rem;
-    color: var(--color-dark-grey-100);
-  }
-
-  &__target:hover &__target__action {
-    border-color: var(--color-dark-grey-200);
-  }
-
-  &__target:hover &__target__action svg {
-    color: var(--color-dark-grey-200);
-  }
-
-  .panel__content--dnd-over &__target__action svg {
-    animation: dnd-bounceup 0.2s alternate ease infinite;
-  }
-
-  &__source-panel,
-  &__expected-result-panel {
-    border-left: 0.1rem solid var(--color-dark-grey-100);
-  }
-
-  &__expected-result-panel__content {
-    position: relative;
-  }
-
-  &__text-editor {
-    padding: 0 !important;
-    position: relative;
-
-    &.panel__content {
-      overflow-y: hidden;
-    }
-  }
-
-  &__generate-result-btn svg {
-    font-size: 1.8rem;
-  }
-}
